Abstracted out a Screen component to be reused across all screens
