+++ frontend/app/src/components/index.js
@@ -3,3 +3,4 @@ export { default as Menu } from "./Menu";
 export { default as Model } from "./Model";
 export { default as Modal } from "./Modal";
 export { default as Chart } from "./Chart";
+export { default as Screen } from "./Screen";
