+++ frontend/app/src/models/index.js
@@ -1,2 +1,2 @@
 export { default as LoadingIndicator } from "./LoadingIndicator";
-export { default as SteelFrame } from "./SteelFrame";
+export { default as SteelFrameModel } from "./SteelFrameModel";
